bup-index: don't warn on lstat failures for excluded paths
Before, when walking the directories, whenever the call to
lstat failed, a warning got raised and bup-index returned with
non-zero exit code. A common case where this happens are fuse
mount points of other users.

This warning could not be suppressed with the exclude* options
because the check for excludes was only performed after the warning
was raised. Now move the check for excludes closer to the lstat call,
inside of _dirlist.

Signed-off-by: Thomas Haller <thom311@gmail.com>