Warn when xattr.get_all fails instead of dying.
xattr.get_all can raise an exception. In that case
bup-index and bup-save should not die, but continue
with a warning.

This failure can also happen on a regular basis.
For example on an encfs filesystem the syscall llistxattr
fails for all broken symlinks. Because of that there
is also a new parameter --suppress-warning-for-xattrs
to convert the warning into a debug1 log message instead.

Signed-off-by: Thomas Haller <thom311@gmail.com>