bugfix fuse-cmd/cache_get to choose the correct cache key
bup-fuse has a function cache_get to lookup a vfs node for a given path.
This function had a bug when accessing paths, whose parent directory is
not yet in the cache itself. For example when accessing /foo/bar without
/foo being cached.

This bug does currently not surface in bup-fuse, because fuse will call getattr
for the parent directories first. This way, bup-fuse populates the cache
in an order, that does not trigger the problem.

With this fix, you can call cache_get for paths in any order and the key
will be chosen correctly.

Signed-off-by: Thomas Haller <thom311@gmail.com>