move fuse/cache_get to its own class vfs.PathLookupIndex
fuse-cmd.py had a method cache_get to cache the lookup for nodes by
path. It does not only cache the requested nodes, but also every parent
node along its path.

This could be useful in other projects as well, so move the method to
its own class in vfs.

Signed-off-by: Thomas Haller <thom311@gmail.com>